Parte 2/Flores/requirements.txt
django
pillow
django-crispy-forms
djangorestframework
django-pwa
